tradingCode/requirements.txt
vectorbtpro
numba>=0.57.0
numpy>=1.24.0
amipy>=0.1.0
pandas>=2.0.0
polars>=0.20.0
pyyaml>=6.0
matplotlib>=3.7.0